Move utility function generate_stream_with_constant_probability to utils.rs and switch to using usize for its initial and final values.
The reason for this change is to be able to use the function for both tests in the lib.rs/tests module and in examples. The architecture remains acyclic at the level of modules but now is cyclic at the level of files.

All tests pass.
